Fixing a bug with my action space in a hacky way for now.
This fixes the large discrepancy between train and test
time and also improves the training in general.

With this, deepQ averages a reward above 1300 at training
time and a reward above 1000 at test time for a 10 x 10 grid.

Compare this to Monte carlo tree search which reaches 1500+
reward.

Of course, so far the environment model is assuming we plant
every cell with something on day1 and harvest all the plants
at once when they are ready (there is no notion of time in
environment that I used for these experiments primrily because
the reward function that we defined doesn't really capture
the implications of time at all).